more reliable handling for proxies in sets, closes #482
